Import a RIF formatted file into a tibble

Reads a file that is stored in RIF format and returns a tibble. The function validates the presence of required fields, and checks whether the variable names in the data match the specified distribution.

import_rif(file)

Arguments

file
Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path) or be a vector of greater than length 1.

Using a value of clipboard() will read from the system clipboard.

Value

An object of class reference, basically a tibble with an attribute called study.

Author

Stef van Buuren 2021

Examples

fn <- system.file("testdata/nl_2009_hgt_male_nl.txt", package = "centile")
myref <- import_rif(fn)