Convert centiles into measurements

Converts centile scores into measurements (KG, CM, ...) using one or more external reference distributions.

p2y(
  p,
  x,
  refcode,
  pkg = "centile",
  verbose = FALSE,
  dec = 3L,
  rule = 1L,
  tail_adjust = FALSE,
  scale = 100,
  ...
)

Arguments

p

A numerical vector with centiles.

x

A vector containing length(y) values of the numerical covariate (typically decimal age or height) at which conversion is desired. A scalar x will be expanded to length(y).

refcode

A character vector with length(y) elements, each of which is the name of a reference. Scalar refcode expands to length(y). Alternatively, refcode may be an object of class reference, typically produced by import_rif(). If NULL the procedure returns NA's.

pkg

The package containing the references in the R/sysdata.dta object. The package needs to be loaded. The default pkg = "centile" searches in the centile package.

verbose

Set to TRUE to turn on warnings

dec

A scalar value indicating the number of decimals used to round the value. The default is 3 decimals.

rule

an integer (of length 1 or 2) describing how interpolation is to take place outside the interval [min(x), max(x)]. If rule is 1 then NAs are returned for such points and if it is 2, the value at the closest data extreme is used. Use, e.g., rule = 2:1, if the left and right side extrapolation should differ.

tail_adjust

Logical. If TRUE the procedure applies the WHO method for tail adjustment under the LMS distribution. We do not recommend this. The default is FALSE.

scale

The maximum of the scale. The default (100) returns for percentiles. Set scale = 1 to obtain probabilities.

...

Not used.

Value

A vector with length(p) elements containing the measurements.

Note

The type of reference distribution is defined by the distribution field of the study attribute in the reference. The function executes transformations specified by the tx field (before calculation of y) and the yt field (after calculating of y), so that the y have the original units even if the reference is based on some transform of y.

Functions p2y() and y2p() functions are the inverse of each other.

Author

Stef van Buuren, 2021

Examples

# Obtain the 10th, 50th and 90th height centile for a boy aged 1 yr on WHO
p2y(c(10, 50, 90), rep(1, 3), "who_2006_hgt_male_")
#> [1] 72.703 75.749 78.794