Load growth reference

This function searches for references stored in the sysdata.dta object in package pkg. If refcode is a reference then

load_reference(
  refcode = NULL,
  element = c("all", "table", "index", "study"),
  pkg = "centile",
  verbose = FALSE
)

Arguments

refcode
String, code of a reference. Only the first element is loaded. Alternatively, if refcode is a reference, the function uses it.

element
Keyword, either "all", "table", "index" or "study". The default is "all".

pkg
The package containing the reference. The package must be loaded and attached. The default pkg = "centile" searches in the home package.

verbose
For more verbose output, set to TRUE.

Value

The return value depends on the element parameter.

element	Return value

"all"	All stored information

"table"	A tibble with reference values

"index"	Numeric vector with index values

"study"	Named character vector with study data

Examples

ref <- load_reference("gc_2019_dsc_male_")